Fixed tooltip, designed and rendered web app editor
Tooltip now has mutation observer if element disappears it no longer remains bugged on the screen.
createNote on the client was bugged so hard for some reason. The "options" were not retrieved ON submit.
Web app moved and developed on its own file
Weather images fixed getting the folder as a random image..

Web app has a working preview and semi working dragging, check it out!!!