MutationObservers! Web app creator upgrade
Way more robust and experimenting with pan zoom library!
It might just turn out great.

Fixed tooltips not disappearing when element is removed rather than mouseout-ed
Fixed ESCStack bugs

Did all that with MutationObservers, something I didn't want to touch but it
just might be fine.

Fixed Ctrl+A not working because of the EscStack.